Beamformer: added assembling tailored gain matrix
Gain matrix in the forward model most likely corresponds to channels that are different from the ones used in the experiment. The could be more (eg. VEOG) but also less (not using all the electrodes). Thus the need to create the gain matrix that corresponds to the channels used in the experiment.